Updates
- Added NetworkStream for buffer writing packets, (Testing)
- Added Spell Cooldowns missing functionality.

+++ Darkages.Server/Network/ServerFormats/ServerFormat3F.cs
@@ -43,9 +43,9 @@ namespace Darkages.Network.ServerFormats
 
         public override void Serialize(NetworkPacketWriter writer)
         {
-            writer.Write(Pane);
-            writer.Write(Slot);
-            writer.Write(Time);
+            writer.Write((byte)Pane);
+            writer.Write((byte)Slot);
+            writer.Write((int)Time);
         }
     }
 }
